feat, refactor : divide REST api (status-service)
- feat: IndexController.java, UserController.java
- refactor : remove empty line
