+++ src/server/status-service/src/main/java/com/example/statusservice/dto/PartnerDto.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class PartnerDto {
-
     private FriendDto requestDto;
     private FriendDto senderOrTargetDto;
 }
